requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
email_validator==2.1.1
et_xmlfile==2.0.0
fastapi==0.110.0
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.26.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflate64==1.0.3
loguru==0.7.2
Mako==1.3.10
MarkupSafe==3.0.2
multivolumefile==0.2.3
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
py7zr==1.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybcj==1.0.6
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyparsing==3.2.3
pyppmd==1.2.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
pyzstd==0.17.0
rarfile==4.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tenacity==8.5.0
texttable==1.7.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.27.1
uvloop==0.21.0
watchdog==4.0.0
watchfiles==1.1.0
websockets==15.0.1
xlrd==2.0.2
